Created LineParser::ParseNextLine() as a start for concurrency
Basic functionality doesn't work for now. It's just a first step to
asynchronous line-parsing.

ParseNextLine will get the line from ifstream, parse it and put into
the unordered_map which will be processed by another thread to check
if year changed between lines and push them further (to a LineWriter
or for emote heuristics).