Fixed some bugs and enabled LineWriter in ProcessMapVarsAsync
ProcessMapVarsAsync now processes until line_map is empty.

If a line can't be parsed, ParseLinesAsynchronously() will add nullptr
to a map instead of just printing an error message. Previous logic
would hang the program because line_map wouldn't get a line with a
certain number and ParseLinesAsynchronously() would wait for it for an
eternity. Now it checks for nullptrs in the line_map.

Removed unnecessary condition_variable, which was never used.

Enabled writing with LineWriter and made some notes about performance.